config: add a new video device
This patch add a new video device "video7" as a first camera
compatible with a interlaced tw9992 sensor.
and the previous video device "video6" is not the first.
the first video device is used by Android default camera app
so it need a modification to set the order for using another video device.

Change-Id: I98670543c426fcb085cb9301d9130059d35d03e7
Signed-off-by: HyejungKwon <cjscld15@nexell.co.kr>